CompositionEngine: Request device composition for the Udfps touched layer
The FOD layers should be always composed by the device to set the
custom zpos bits to the kernel. By default all layers are set to
device composition and the proprietary libsdmextension.so moves
some to client composition. Unfortunately that also affects the
Udfps touched layer. After this commit the layer below the Udfps
touched layer is forced to be client composition. For unknown
reasons libsdmextension will not change the Udfps touched layer
to client composition anymore.

tests:
- Run 'adb shell dumpsys SurfaceFlinger' and ensure that the Udfps
  touched layer is composed by the device
- Ensure that the Udfps touched layer correctly sets the zpos bits
  on the kernel while / after using WFD

Signed-off-by: daniml3 <danimoral1001@gmail.com>
Signed-off-by: Arian <arian.kulmer@web.de>
Change-Id: I8aeb98d18557ad4e971eaba74700ceb3058273ab
Signed-off-by: Noah Anleitner <noah.anleitner@halogenos.org>
Signed-off-by: Simão Gomes Viana <devel@superboring.dev>